refactor(cli): run under Node, extract sync progress, improve post-sync hints
- Use a Node shebang so npx notion matches engines.node.
- Extract SyncProgressRenderer for TTY vs non-TTY progress lines.
- Delegate init/template helpers to config/init; import findConfigFile directly.
- Use PACKAGE_RUNTIME_CONSTANTS for Notion API version in fetch helpers.
- Print a minimal notion entrypoint example after sync; simplify DB/agent summaries.

Made-with: Cursor
